# backend/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Team Manager API
  description: >-
    REST API for the Teamverwaltung sports-club management application.


    Every endpoint may return `429 Too Many Requests` (global per-IP rate
    limit, `RATE_LIMIT_RPS`) or `413 Payload Too Large` (request body cap,
    4 MB by default) — see the `TooManyRequests`/`PayloadTooLarge` reusable
    responses. These aren't repeated on every operation below to keep the
    spec readable; `/auth/login` documents its own stricter rate limit
    explicitly since it differs from the global default.
  version: "1.0.0"
  contact:
    name: Team Manager
    url: https://github.com/yoadey/team-manager
servers:
  - url: /api/v1
    description: Current server
security:
  - bearerAuth: []

# ─── Tags ────────────────────────────────────────────────────────────────────
...
paths:
...
  /teams/{teamId}/finances:
    parameters:
      - $ref: "#/components/parameters/teamId"
    get:
      tags: [finances]
      operationId: getFinanceOverview
      summary: Get full financial overview
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceOverview"
  /teams/{teamId}/finances/transactions:
    parameters:
      - $ref: "#/components/parameters/teamId"
    post:
      tags: [finances]
      operationId: createTransaction
      summary: Add income or expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /teams/{teamId}/finances/transactions/{transactionId}:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - name: transactionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [finances]
      operationId: updateTransaction
      summary: Update transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      tags: [finances]
      operationId: deleteTransaction
      summary: Delete transaction
      responses:
        "204":
          description: No Content
  /teams/{teamId}/finances/penalties:
    parameters:
      - $ref: "#/components/parameters/teamId"
    post:
      tags: [finances]
      operationId: createPenalty
      summary: Create penalty template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePenaltyRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Penalty"
  /teams/{teamId}/finances/penalties/{penaltyId}:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - name: penaltyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [finances]
      operationId: updatePenalty
      summary: Update penalty template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePenaltyRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Penalty"
    delete:
      tags: [finances]
      operationId: deletePenalty
      summary: Delete penalty template and all assignments
      responses:
        "204":
          description: No Content
  /teams/{teamId}/finances/penalty-assignments:
    parameters:
      - $ref: "#/components/parameters/teamId"
    post:
      tags: [finances]
      operationId: createPenaltyAssignment
      summary: Assign penalty to member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePenaltyAssignmentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenaltyAssignment"
  /teams/{teamId}/finances/penalty-assignments/{assignmentId}:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - name: assignmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [finances]
      operationId: deletePenaltyAssignment
      summary: Remove penalty assignment
      responses:
        "204":
          description: No Content
  /teams/{teamId}/finances/penalty-assignments/{assignmentId}/toggle-paid:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - name: assignmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [finances]
      operationId: togglePenaltyPaid
      summary: Toggle paid status of penalty assignment
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenaltyAssignment"
  /teams/{teamId}/finances/contributions/{contributionId}:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - name: contributionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [finances]
      operationId: updateContribution
      summary: Update contribution amount or label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContributionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"
  /teams/{teamId}/finances/contributions/{contributionId}/toggle:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - name: contributionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [finances]
      operationId: toggleContribution
      summary: Toggle contribution paid/open
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"

  # ── Stats ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    teamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TransactionType:
      type: string
      enum: [income, expense]
    ContributionStatus:
      type: string
      enum: [paid, open]
...
    Transaction:
      type: object
      required: [id, teamId, type, title, amount, date]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/TransactionType"
        title:
          type: string
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
        date:
          type: string
          format: date
        category:
          type: string
    CreateTransactionRequest:
      type: object
      required: [type, title, amount]
      properties:
        type:
          $ref: "#/components/schemas/TransactionType"
        title:
          type: string
          maxLength: 255
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
          minimum: 1
          maximum: 100000000
        category:
          type: string
          maxLength: 255
    UpdateTransactionRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/TransactionType"
        title:
          type: string
          maxLength: 255
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
          minimum: 1
          maximum: 100000000
        category:
          type: string
          maxLength: 255
    Penalty:
      type: object
      required: [id, teamId, label, amount]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        label:
          type: string
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
    CreatePenaltyRequest:
      type: object
      required: [label, amount]
      properties:
        label:
          type: string
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
          minimum: 1
          maximum: 100000000
    UpdatePenaltyRequest:
      type: object
      properties:
        label:
          type: string
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
          minimum: 1
          maximum: 100000000
    PenaltyAssignment:
      type: object
      required: [id, teamId, userId, penaltyId, paid, date]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        penaltyId:
          type: string
          format: uuid
        paid:
          type: boolean
        date:
          type: string
          format: date
        memberName:
          type: string
        memberAvatarColor:
          type: string
        hasPhoto:
          type: boolean
        label:
          type: string
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
    CreatePenaltyAssignmentRequest:
      type: object
      required: [userId, penaltyId]
      properties:
        userId:
          type: string
          format: uuid
        penaltyId:
          type: string
          format: uuid
    OpenPenalty:
      type: object
      required: [userId, name, avatarColor, amount]
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        avatarColor:
          type: string
        hasPhoto:
          type: boolean
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
    Contribution:
      type: object
      required: [id, teamId, userId, month, amount, status]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        month:
          type: string
          description: "YYYY-MM"
        label:
          type: string
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
        status:
          $ref: "#/components/schemas/ContributionStatus"
        memberName:
          type: string
        memberAvatarColor:
          type: string
        hasPhoto:
          type: boolean
    UpdateContributionRequest:
      type: object
      properties:
        label:
          type: string
        amount:
          type: integer
          format: int64
          description: Amount in cents (e.g. 1050 = 10.50)
          minimum: 1
          maximum: 100000000
    FinanceOverview:
      type: object
      required: [balance, income, expense, transactions, penalties, assignments, openPenalties, openPenaltySum, contributions, contribOpen]
      properties:
        balance:
          type: integer
          format: int64
          description: Balance in cents (income minus expense)
        income:
          type: integer
          format: int64
          description: Total income in cents
        expense:
          type: integer
          format: int64
          description: Total expense in cents
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        penalties:
          type: array
          items:
            $ref: "#/components/schemas/Penalty"
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/PenaltyAssignment"
        openPenalties:
          type: array
          items:
            $ref: "#/components/schemas/OpenPenalty"
        openPenaltySum:
          type: integer
          format: int64
          description: Sum of all open (unpaid) penalties across all members, in cents
        contributions:
          type: array
          items:
            $ref: "#/components/schemas/Contribution"
        contribOpen:
          type: integer

    # ── Stats ────────────────────────────────────────────────────────────────
...